// src/constants/enums.ts
export enum EnumStatus {
  NEW = 'NEW',
  POSTED = 'POSTED',
  ASSIGNED = 'ASSIGNED',
  SHIPPED = 'SHIPPED',
}
export enum EnumState {
  ROUTE_TO = 'En route to Pick Up',
  ARRIVED = 'Arrived to Pick Up',
  DELIVERY = 'En route to delivery',
  FINISH = 'Arrived to delivery',
}
export enum EnumRole {
  SHIPPER = 'SHIPPER',
  DRIVER = 'DRIVER',
}
export enum EnumTruckType {
  SPRINTER = 'SPRINTER',
  SMALL_STRAIGHT = 'SMALL STRAIGHT',
  LARGE_STRAIGHT = 'LARGE STRAIGHT',
}
export enum EnumTruckStatus {
  IS = 'IS',
  OL = 'OL',
}